-AddWord || + WordConfiguration
change AddWord name to WordConfiguration name
