+++ src/main/java/com/example/translatetest1/Main.java
@@ -6,10 +6,16 @@ import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 import java.io.IOException;
+import java.sql.SQLException;
 
 public class Main extends Application {
     @Override
-    public void start(Stage stage) throws IOException {
+    public void init() throws Exception {
+        DataBaseManager.init();
+    }
+
+    @Override
+    public void start(Stage stage) throws IOException, SQLException {
         FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("HelloApplication.fxml"));
         Scene scene = new Scene(fxmlLoader.load(), 320, 240);
         System.out.println("java version: " + System.getProperty("java.version"));
@@ -17,5 +23,7 @@ public class Main extends Application {
         stage.setScene(scene);
         stage.show();
     }
-    public static void main(String [] args){launch();}
+    public static void main(String [] args) {
+        launch();
+    }
 }
