+++ src/main/java/com/example/translatetest1/MyDictionary.java
@@ -1,6 +1,7 @@
 package com.example.translatetest1;
 
 import Singleton.Singleton;
+import javafx.scene.control.Alert;
 
 import java.util.Dictionary;
 import java.util.*;
@@ -8,8 +9,20 @@ import java.util.*;
   class MyDictionary extends Singleton<MyDictionary> {
     public static Dictionary<String, String> dic = new Hashtable<>();
     public static void addWordtodic(String word, String definition) {
-        if(!isWordInDic(word)) {
+        if (isWordInDic(word)) {
             dic.put(word, definition);
+            DataBaseManager.addWordtoSQL(word, definition);
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setTitle("Information dialog");
+            alert.setContentText("Từ " + word + " đã được thêm thành công!");
+            alert.setHeaderText("Thông báo");
+            alert.showAndWait();
+        } else {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+            alert.setTitle("Information dialog");
+            alert.setContentText("Từ " + word + " đã tồn tại trong từ điển!");
+            alert.setHeaderText("Thông báo");
+            alert.showAndWait();
         }
     }
