+++ src/main/resources/com/example/translatetest1/WordConfiguration.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane fx:id="AddPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.translatetest1.AddWord">
+<AnchorPane fx:id="AddPane" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/20.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.translatetest1.WordConfiguration">
    <children>
       <Label layoutX="14.0" layoutY="5.0" text="Nhập từ cần thêm" />
       <TextField fx:id="addbox1" layoutX="1.0" layoutY="46.0" prefHeight="26.0" prefWidth="522.0" />
